Update neighbor_finder.R
added an error note if the gene of interest is not in the matrix. Before, the function would calculate first (which could take some time) and then give an error if the gene is not present. Now, this is checked before everything else.

removed CDKNA1 as the default gene (if this is accidentally forgotten by the user it is better the function complains, than to spend time calculating something the user does not want)

added some function description